[feat] 캐러셀 1차 완성
